+++ itsaviya/server/controllers/blog/blog-controller.js
@@ -1,14 +1,12 @@
-const { Blog } = require("../models/blogModel.js");
+const { Blog } = require("../../models/blogModel.js");
 
 const getBlogByTitle = async (req, res) => {
-  console.log("getting blog by title");
   const name = req.params.title;
   const blog = await Blog.findOne({ name: name });
   res.json({ blog });
 };
 
 const getAllBlogs = async (req, res) => {
-  console.log("getAllBlogs");
   const blogs = await Blog.find({}); //need to filter when more fields will be added
   res.json(blogs);
 };
