+++ itsaviya/server/controllers/user/email.js
@@ -1,0 +1,11 @@
+const nodemailer = require("nodemailer");
+
+const transporter = nodemailer.createTransport({
+  service: "gmail",
+  auth: {
+    user: "testkipi233@gmail.com",
+    pass: "gbbipumuvknsnljw",
+  },
+});
+
+module.exports = { transporter };
