+++ itsaviya/server/controllers/user/refreshToken-controller.js
@@ -1,4 +1,4 @@
-const { User } = require("../models/userModel.js");
+const { User } = require("../../models/userModel.js");
 
 const jwt = require("jsonwebtoken");
 require("dotenv").config();
